organize Datapack module
