+++ src/main/java/com/voltskiya/misc/datapack/ModuleDatapack.java
@@ -1,6 +1,8 @@
 package com.voltskiya.misc.datapack;
 
 import com.voltskiya.lib.AbstractModule;
+import com.voltskiya.misc.datapack.damage.ScoreboardListener;
+import com.voltskiya.misc.datapack.reload.CommandDatapackReload;
 
 public class ModuleDatapack extends AbstractModule {
